Refactor resizer
